+++ ArtOfIllusion/src/artofillusion/object/SpotLight.java
@@ -352,7 +352,7 @@ public class SpotLight extends Light
     else if (index == 4)
       decayRate = ((Double) value).floatValue();
     else if (index == 5)
-      radius = ((Double) value).doubleValue();
+      radius = ((Double) value);
     else if (index == 6)
     {
       Object values[] = PROPERTIES[index].getAllowedValues();
